+++ packages/react/Badge/src/badge.types.ts
@@ -10,4 +10,6 @@ export interface BadgeProps extends HTMLAttributes<HTMLElement> {
   appearance?: "primary" | "success" | "warning" | "danger" | "neutral";
   /** Total items to be displayed without badge */
   count: number | string;
+  /** Badge appearance color scheme */
+  theme?: "surface" | "light";
 }