+++ packages/react/Skeleton/src/Skeleton.tsx
@@ -23,4 +23,5 @@ const Skeleton: React.FC<SkeletonProps> = ({
   />
 );
 
+Skeleton.displayName = "Skeleton";
 export { Skeleton };